add backwards/forwards compatibility test.
